reorganized file paths and namespaces
