+++ 0.Core/Core/ExtensionHelpers/ViewModelExtensions.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Text;
 using System.Web;
 
-namespace ApiServer.Tests
+namespace Core.ExtensionHelpers
 {
     public static class ViewModelExtensions
     {
