+++ 1.Domain/Domain.UnitTests/Services/SkillTypeServiceTest.cs
@@ -3,11 +3,11 @@ using System.Linq;
 using AutoMapper;
 using Core;
 using Core.Persistance;
+using Core.Testing;
 using Domain.Model;
 using Domain.Services.Contracts.SkillType;
 using Domain.Services.Impl.Services;
 using Domain.Services.Impl.Validators.SkillType;
-using Domain.Services.Tests;
 using FluentValidation;
 using FluentValidation.Results;
 using Moq;
