+++ 1.Domain/Domain.UnitTests/Services/TaskServiceTest.cs
@@ -3,12 +3,12 @@ using System.Linq;
 using AutoMapper;
 using Core;
 using Core.Persistance;
+using Core.Testing;
 using Domain.Model;
 using Domain.Services.Contracts.Task;
 using Domain.Services.Contracts.TaskItem;
 using Domain.Services.Impl.Services;
 using Domain.Services.Impl.Validators.Task;
-using Domain.Services.Tests;
 using FluentValidation;
 using FluentValidation.Results;
 using Moq;
