+++ 1.Domain/Domain.UnitTests/Validators/Candidate/UpdateCandidateContractValidatorTester.cs
@@ -6,7 +6,7 @@ using FluentValidation.TestHelper;
 using System;
 using Xunit;
 
-namespace Domain.Services.Impl.UnitTests.ValidatorTester.Candidate
+namespace Domain.Services.Impl.UnitTests.Validators.Candidate
 {
     public class UpdateCandidateContractValidatorTester
     {
