+++ 1.Domain/Domain.UnitTests/Validators/Community/UpdateCommunityContractValidatorTester.cs
@@ -3,15 +3,15 @@ using FluentValidation.TestHelper;
 using System;
 using Xunit;
 
-namespace Domain.Services.Impl.UnitTests.ValidatorTester.Community
+namespace Domain.Services.Impl.UnitTests.Validators.Community
 {
-    public class CreateCommunityContractValidatorTester
+    public class UpdateCommunityContractValidatorTester
     {
-        private readonly CreateCommunityContractValidator _validator;
+        private readonly UpdateCommunityContractValidator _validator;
 
-        public CreateCommunityContractValidatorTester()
+        public UpdateCommunityContractValidatorTester()
         {
-            _validator = new CreateCommunityContractValidator();
+            _validator = new UpdateCommunityContractValidator();
         }
 
         [Fact(DisplayName = "Verify that throws error when Name string is null")]
@@ -30,6 +30,12 @@ namespace Domain.Services.Impl.UnitTests.ValidatorTester.Community
         public void Should_Have_Error_When_Description_Is_Null()
         {
             _validator.ShouldHaveValidationErrorFor(CommunityContract => CommunityContract.Description, (string) null);
+        }
+
+        [Fact(DisplayName = "Verify that throws error when Description string is Empty")]
+        public void Should_Have_Error_When_Description_Is_Empty()
+        {
+            _validator.ShouldHaveValidationErrorFor(CommunityContract => CommunityContract.Description, String.Empty);
         }
     }
 }
