+++ 4.Clients/ApiServer.UnitTests/Controllers/TasksControllerTest.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Generic;
 using ApiServer.Contracts.Task;
 using ApiServer.Controllers;
-using ApiServer.Tests;
 using AutoMapper;
 using Core;
+using Core.ExtensionHelpers;
 using Domain.Services.Contracts.Task;
 using Domain.Services.Interfaces.Services;
 using Microsoft.AspNetCore.Mvc;
